Move identity service to sentry.identity
