+++ src/sentry/integrations/base.py
@@ -15,6 +15,8 @@ from rest_framework.exceptions import NotFound
 from sentry import audit_log, features
 from sentry.constants import ObjectStatus
 from sentry.exceptions import InvalidIdentity
+from sentry.identity.services.identity import identity_service
+from sentry.identity.services.identity.model import RpcIdentity
 from sentry.integrations.notify_disable import notify_disable
 from sentry.integrations.request_buffer import IntegrationRequestBuffer
 from sentry.models.identity import Identity
@@ -23,8 +25,6 @@ from sentry.models.integrations.integration import Integration
 from sentry.models.team import Team
 from sentry.pipeline import PipelineProvider
 from sentry.pipeline.views.base import PipelineView
-from sentry.services.hybrid_cloud.identity import identity_service
-from sentry.services.hybrid_cloud.identity.model import RpcIdentity
 from sentry.services.hybrid_cloud.organization import (
     RpcOrganization,
     RpcOrganizationSummary,
