+++ src/sentry/integrations/discord/requests/base.py
@@ -9,10 +9,10 @@ from rest_framework import status
 from rest_framework.request import Request
 
 from sentry import options
+from sentry.identity.services.identity import RpcIdentityProvider
+from sentry.identity.services.identity.model import RpcIdentity
+from sentry.identity.services.identity.service import identity_service
 from sentry.integrations.services.integration import RpcIntegration, integration_service
-from sentry.services.hybrid_cloud.identity import RpcIdentityProvider
-from sentry.services.hybrid_cloud.identity.model import RpcIdentity
-from sentry.services.hybrid_cloud.identity.service import identity_service
 from sentry.services.hybrid_cloud.user.model import RpcUser
 from sentry.services.hybrid_cloud.user.service import user_service
