+++ src/sentry/integrations/notifications.py
@@ -5,13 +5,13 @@ from collections.abc import Iterable, Mapping, MutableMapping
 from typing import Any
 
 from sentry.constants import ObjectStatus
+from sentry.identity.services.identity import identity_service
 from sentry.integrations.services.integration import RpcIntegration, integration_service
 from sentry.integrations.types import EXTERNAL_PROVIDERS, ExternalProviders
 from sentry.models.integrations.external_actor import ExternalActor
 from sentry.models.organization import Organization
 from sentry.models.team import Team
 from sentry.notifications.notifications.base import BaseNotification
-from sentry.services.hybrid_cloud.identity import identity_service
 from sentry.services.hybrid_cloud.user import RpcUser
 from sentry.types.actor import Actor
