+++ src/sentry/integrations/services/integration/impl.py
@@ -485,7 +485,7 @@ class DatabaseBackedIntegrationService(IntegrationService):
         identity_external_id: str | None = None,
         identity_provider_external_id: str | None = None,
     ) -> RpcIntegrationIdentityContext:
-        from sentry.services.hybrid_cloud.identity.service import identity_service
+        from sentry.identity.services.identity.service import identity_service
         from sentry.services.hybrid_cloud.user.service import user_service
 
         integration = self.get_integration(
