+++ src/sentry/integrations/slack/requests/base.py
@@ -9,9 +9,9 @@ from rest_framework.request import Request
 from slack_sdk.signature import SignatureVerifier
 
 from sentry import options
+from sentry.identity.services.identity import RpcIdentity, identity_service
+from sentry.identity.services.identity.model import RpcIdentityProvider
 from sentry.integrations.services.integration import RpcIntegration, integration_service
-from sentry.services.hybrid_cloud.identity import RpcIdentity, identity_service
-from sentry.services.hybrid_cloud.identity.model import RpcIdentityProvider
 from sentry.services.hybrid_cloud.user import RpcUser
 from sentry.services.hybrid_cloud.user.service import user_service
 from sentry.utils.safe import get_path
