+++ src/sentry/receivers/outbox/control.py
@@ -15,6 +15,7 @@ from typing import Any
 from django.dispatch import receiver
 
 from sentry.hybridcloud.rpc.caching import region_caching_service
+from sentry.issues.services.issue import issue_service
 from sentry.models.apiapplication import ApiApplication
 from sentry.models.integrations.integration import Integration
 from sentry.models.integrations.sentry_app import SentryApp
@@ -23,7 +24,6 @@ from sentry.models.organizationmapping import OrganizationMapping
 from sentry.models.outbox import OutboxCategory, process_control_outbox
 from sentry.receivers.outbox import maybe_process_tombstone
 from sentry.sentry_apps.services.app.service import get_by_application_id, get_installation
-from sentry.services.hybrid_cloud.issue import issue_service
 from sentry.services.hybrid_cloud.organization import RpcOrganizationSignal, organization_service
 
 logger = logging.getLogger(__name__)
