+++ tests/sentry/integrations/test_base.py
@@ -1,5 +1,5 @@
+from sentry.identity.services.identity.serial import serialize_identity
 from sentry.integrations.base import IntegrationInstallation
-from sentry.services.hybrid_cloud.identity.serial import serialize_identity
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import all_silo_test, assume_test_silo_mode
